+++ packages/field/index.ts
@@ -66,6 +66,20 @@ VantComponent({
     showClear: false,
   },
 
+  watch: {
+    value(this: WechatMiniprogram.Component.TrivialInstance, value) {
+      if (value !== this.value) {
+        this.setData({ innerValue: value });
+        this.value = value;
+
+        this.setShowClear();
+      }
+    },
+    clearTrigger() {
+      this.setShowClear();
+    },
+  },
+
   created() {
     this.value = this.data.value;
     this.setData({ innerValue: this.value });
@@ -164,6 +178,10 @@ VantComponent({
       this.$emit('keyboardheightchange', event.detail);
     },
 
+    onBindNicknameReview(event) {
+      this.$emit('nicknamereview', event.detail);
+    },
+
     emitChange(detail: InputDetails) {
       const { extraEventParams } = this.data;
 
@@ -200,7 +218,7 @@ VantComponent({
         showClear = hasValue && trigger;
       }
 
-      this.setData({ showClear });
+      this.setView({ showClear });
     },
 
     noop() {},