+++ packages/field/props.ts
@@ -1,13 +1,5 @@
 export const commonProps: WechatMiniprogram.Component.PropertyOption = {
-  value: {
-    type: String,
-    observer(this: WechatMiniprogram.Component.TrivialInstance, value) {
-      if (value !== this.value) {
-        this.setData({ innerValue: value });
-        this.value = value;
-      }
-    },
-  },
+  value: String,
   placeholder: String,
   placeholderStyle: String,
   placeholderClass: String,