+++ packages/mixins/basic.ts
@@ -13,5 +13,28 @@ export const basic = Behavior({
 
       return new Promise((resolve) => wx.nextTick(resolve));
     },
+
+    // high performance setData
+    setView(
+      this: WechatMiniprogram.Component.TrivialInstance,
+      data: Record<string, any>,
+      callback?: () => void
+    ) {
+      const target: Record<string, any> = {};
+      let hasChange = false;
+
+      Object.keys(data).forEach((key) => {
+        if (data[key] !== this.data[key]) {
+          target[key] = data[key];
+          hasChange = true;
+        }
+      });
+
+      if (hasChange) {
+        return this.setData(target, callback);
+      }
+
+      return callback && callback();
+    },
   },
 });