+++ packages/mixins/transition.ts
@@ -23,7 +23,6 @@ export function transition(showDefaultValue: boolean) {
       duration: {
         type: null,
         value: 300,
-        observer: 'observeDuration',
       },
       name: {
         type: String,
@@ -49,89 +48,101 @@ export function transition(showDefaultValue: boolean) {
           return;
         }
 
-        value ? this.enter() : this.leave();
+        value ? this.enureEnter() : this.enureLeave();
       },
 
-      enter() {
-        if (this.enterFinishedPromise) return;
-        this.enterFinishedPromise = new Promise((resolve) => {
-          const { duration, name } = this.data;
-          const classNames = getClassNames(name);
-          const currentDuration = isObj(duration) ? duration.enter : duration;
+      enureEnter() {
+        if (this.enterPromise) return;
 
-          if (this.status === 'enter') {
+        this.enterPromise = new Promise((resolve) => this.enter(resolve));
+      },
+
+      enureLeave() {
+        const { enterPromise } = this;
+
+        if (!enterPromise) return;
+
+        enterPromise
+          .then(() => new Promise((resolve) => this.leave(resolve)))
+          .then(() => {
+            this.enterPromise = null;
+          });
+      },
+
+      enter(resolve: () => void) {
+        const { duration, name } = this.data;
+        const classNames = getClassNames(name);
+        const currentDuration = isObj(duration) ? duration.enter : duration;
+
+        if (this.status === 'enter') {
+          return;
+        }
+
+        this.status = 'enter';
+        this.$emit('before-enter');
+
+        requestAnimationFrame(() => {
+          if (this.status !== 'enter') {
             return;
           }
 
-          this.status = 'enter';
-          this.$emit('before-enter');
+          this.$emit('enter');
+
+          this.setData({
+            inited: true,
+            display: true,
+            classes: classNames.enter,
+            currentDuration,
+          });
 
           requestAnimationFrame(() => {
             if (this.status !== 'enter') {
               return;
             }
 
-            this.$emit('enter');
-
-            this.setData({
-              inited: true,
-              display: true,
-              classes: classNames.enter,
-              currentDuration,
-            });
-
-            requestAnimationFrame(() => {
-              if (this.status !== 'enter') {
-                return;
-              }
-
-              this.transitionEnded = false;
-              this.setData({ classes: classNames['enter-to'] });
-              resolve();
-            });
+            this.transitionEnded = false;
+            this.setData({ classes: classNames['enter-to'] });
+            resolve();
           });
         });
       },
 
-      leave() {
-        if (!this.enterFinishedPromise) return;
-        this.enterFinishedPromise.then(() => {
-          if (!this.data.display) {
+      leave(resolve: () => void) {
+        if (!this.data.display) {
+          return;
+        }
+
+        const { duration, name } = this.data;
+        const classNames = getClassNames(name);
+        const currentDuration = isObj(duration) ? duration.leave : duration;
+
+        this.status = 'leave';
+        this.$emit('before-leave');
+
+        requestAnimationFrame(() => {
+          if (this.status !== 'leave') {
             return;
           }
 
-          const { duration, name } = this.data;
-          const classNames = getClassNames(name);
-          const currentDuration = isObj(duration) ? duration.leave : duration;
+          this.$emit('leave');
 
-          this.status = 'leave';
-          this.$emit('before-leave');
+          this.setData({
+            classes: classNames.leave,
+            currentDuration,
+          });
 
           requestAnimationFrame(() => {
             if (this.status !== 'leave') {
               return;
             }
 
-            this.$emit('leave');
+            this.transitionEnded = false;
+            setTimeout(() => {
+              this.onTransitionEnd();
+              resolve();
+            }, currentDuration);
 
-            this.setData({
-              classes: classNames.leave,
-              currentDuration,
-            });
-
-            requestAnimationFrame(() => {
-              if (this.status !== 'leave') {
-                return;
-              }
-
-              this.transitionEnded = false;
-              setTimeout(() => {
-                this.onTransitionEnd();
-                this.enterFinishedPromise = null;
-              }, currentDuration);
-
-              this.setData({ classes: classNames['leave-to'] });
-            });
+            this.setData({ classes: classNames['leave-to'] });
           });
         });
       },