+++ packages/uploader/index.ts
@@ -1,7 +1,7 @@
 import { VantComponent } from '../common/component';
-import { isImageFile, chooseFile, isVideoFile, File } from './utils';
-import { imageProps, videoProps, mediaProps, messageFileProps } from './shared';
 import { isBoolean, isPromise } from '../common/validator';
+import { imageProps, mediaProps, messageFileProps, videoProps } from './shared';
+import { chooseFile, File, isImageFile, isVideoFile } from './utils';
 
 VantComponent({
   props: {
@@ -218,6 +218,8 @@ VantComponent({
     },
 
     onPreviewFile(event: WechatMiniprogram.TouchEvent) {
+      if (!this.data.previewFile) return;
+
       const { index } = event.currentTarget.dataset;
 
       wx.openDocument({