+++ packages/uploader/shared.ts
@@ -32,6 +32,10 @@ export const videoProps = {
     type: String,
     value: 'back',
   },
+  referrerPolicy: {
+    type: String,
+    value: 'no-referrer',
+  },
 };
 
 // props for media
@@ -57,4 +61,8 @@ export const mediaProps = {
 // props for file
 export const messageFileProps = {
   extension: null,
+  previewFile: {
+    type: Boolean,
+    value: true,
+  },
 };