+++ rsbuild.config.ts
@@ -1,0 +1,7 @@
+export default {
+  server: {
+    proxy: {
+      '/vant/v2': 'https://vant-ui.github.io',
+    },
+  },
+};